Refactor - move `isSupportEmailNotificationAllowed` from `infrastructure/applications` to  to a helper file.
Test Suites: 165 passed, 165 total
Tests:       928 passed, 928 total
Snapshots:   0 total
Time:        4.232 s
