+++ src/infrastructure/applications/static.js
@@ -2,11 +2,6 @@ const createService = async () => {
   return Promise.resolve();
 };
 
-const isSupportEmailNotificationAllowed = async () => {
-  return Promise.resolve([]);
-};
-
 module.exports = {
   createService,
-  isSupportEmailNotificationAllowed,
 };
