+++ test/app/users/postConfirmAddService.test.js
@@ -6,13 +6,9 @@ jest.mock("./../../../src/infrastructure/logger", () =>
 );
 jest.mock("./../../../src/infrastructure/organisations");
 jest.mock("login.dfe.api-client/invitations");
-jest.mock("./../../../src/infrastructure/applications", () => {
-  return {
-    isSupportEmailNotificationAllowed: jest.fn(),
-  };
-});
 jest.mock("../../../src/app/services/utils", () => ({
   getAllServices: jest.fn(),
+  isSupportEmailNotificationAllowed: jest.fn(),
 }));
 
 jest.mock("./../../../src/infrastructure/access", () => {
@@ -22,7 +18,6 @@ jest.mock("./../../../src/infrastructure/access", () => {
     updateUserService: jest.fn(),
   };
 });
-
 const { getRequestMock, getResponseMock } = require("./../../utils");
 const {
   listRolesOfService,
@@ -35,7 +30,7 @@ const {
 } = require("login.dfe.api-client/invitations");
 const {
   isSupportEmailNotificationAllowed,
-} = require("./../../../src/infrastructure/applications");
+} = require("../../../src/app/services/utils");
 const { getAllServices } = require("../../../src/app/services/utils");
 const {
   getUserOrganisations,
