+++ test/app/users/postDeleteOrganisation.test.js
@@ -19,6 +19,7 @@ jest.mock("./../../../src/infrastructure/search", () => {
 });
 
 jest.mock("./../../../src/app/users/utils");
+jest.mock("login.dfe.api-client/services");
 
 jest.mock("login.dfe.jobs-client");
 const { NotificationClient } = require("login.dfe.jobs-client");
